+++ resources/js/forms/components/fields/number.vue
@@ -1,22 +1,30 @@
 <template>
   <div class="relative">
-    <textarea
+    <input
+      type="number"
+      min="1"
+      step="1"
       :value="modelValue"
       @input="$emit('update:modelValue', $event.target.value)"
       @focus="$emit('update:error', '')"
+      :placeholder="placeholder"
       :class="[
         { '!border-raspberry': error },
-        classes
-        ]">
-    </textarea>
+      ]"
+    >
     <Error :error="error" />
   </div>
 </template>
 
 <script setup>
 import Error from './error.vue';
+
 const props = defineProps({
   modelValue: {
+    type: [String, Number],
+    default: ''
+  },
+  placeholder: {
     type: String,
     default: ''
   },
@@ -24,10 +32,6 @@ const props = defineProps({
     type: String,
     default: ''
   },
-  classes: {
-    type: String,
-    default: ''
-  }
 });
 
 defineEmits(['update:modelValue', 'update:error']);