+++ CHANGELOG.md
@@ -1,3 +1,4 @@
-27.06.2025 – Bestellformular für Publikationen
-
+## 27.06.2025 – Bestellformular für Publikationen
 - Set link in /views/partials/publication/item.antlers.html
+- Create Page "Bestellung Arbeitshilfen & Publikationen" as child page of "Arbeitshilfen & Publikationen"
+- Integrate "Bestellformular Publikationen"